// src/types/guildScheduledEvents/frequency.ts
/**
 * Guild Scheduled Event Recurrency Rule - Frequency
 *
 * @see [Guild Scheduled Event Recurrency Rule - Frequency](https://discord.com/developers/docs/resources/guild-scheduled-event#guild-scheduled-event-recurrence-rule-object-guild-scheduled-event-recurrence-rule-frequency)
 *
 * @enum
 */
export enum Frequency {
	/**
	 * Yearly
	 */
	YEARLY = 0,

	/**
	 * Monthly
	 */
	MONTHLY = 1,

	/**
	 * Weekly
	 */
	WEEKLY = 2,

	/**
	 * Daily
	 */
	DAILY = 3
}