// src/types/guildScheduledEvents/weekday.ts
/**
 * Guild Scheduled Event Recurrency Rule - Weekday
 *
 * @see [Guild Scheduled Event Recurrency Rule - Weekday](https://discord.com/developers/docs/resources/guild-scheduled-event#guild-scheduled-event-recurrence-rule-object-guild-scheduled-event-recurrence-rule-weekday)
 *
 * @enum
 */
export enum Weekday {
	/**
	 * Monday
	 */
	MONDAY = 0,

	/**
	 * Tuesday
	 */
	TUESDAY = 1,

	/**
	 * Wednesday
	 */
	WEDNESDAY = 2,

	/**
	 * Thursday
	 */
	THURSDAY = 3,

	/**
	 * Friday
	 */
	FRIDAY = 4,

	/**
	 * Saturday
	 */
	SATURDAY = 5,

	/**
	 * Sunday
	 */
	SUNDAY = 6
}