// src/types/guildScheduledEvents/month.ts
/**
 * Guild Scheduled Event Recurrency Rule - Month
 *
 * @see [Guild Scheduled Event Recurrency Rule - Month](https://discord.com/developers/docs/resources/guild-scheduled-event#guild-scheduled-event-recurrence-rule-object-guild-scheduled-event-recurrence-rule-month)
 *
 * @enum
 */
export enum Month {
	/**
	 * January
	 */
	JANUARY = 1,

	/**
	 * February
	 */
	FEBRUARY = 2,

	/**
	 * March
	 */
	MARCH = 3,

	/**
	 * April
	 */
	APRIL = 4,

	/**
	 * May
	 */
	MAY = 5,

	/**
	 * June
	 */
	JUNE = 6,

	/**
	 * July
	 */
	JULY = 7,

	/**
	 * August
	 */
	AUGUST = 8,

	/**
	 * September
	 */
	SEPTEMBER = 9,

	/**
	 * October
	 */
	OCTOBER = 10,

	/**
	 * November
	 */
	NOVEMBER = 11,

	/**
	 * December
	 */
	DECEMBER = 12
}